Fixed a defect in which ObjectDoesNotExists would not be thrown for some invalid entries in ArchiveRecordStores.
lower_bound() uses STL comparators to do comparisons and the default comparator
for map is less<>, meaning it could return a valid iterator match when keys do
not match if a key exists that might be considered "close."

To resolve this for now, switch to find(), which only returns exact key matches.  
It will be advantageous to return to using lower_bound() and explicitly set the
comparator to equal_to<string> when there is more time for testing, due to the
speed difference between the two standard algorithms.

[git-p4: depot-paths = "//depot/projects/EVALUATION/Main/": change = 2560]