Fix bug where ListRecordStores could not be modified on NFS devices.
Instead of directly modifying KeyList.txt, a temporary copy was made
in /tmp to prevent a corrupted RS if rstool crashed while modifying
the list. The temporary KeyList.txt was then moved from /tmp to the
RecordStore with rename(2). Unfortunately, rename(2) cannot be safely
used across devices. Fix the issue by creating the temporary
KeyList.txt directly in the RecordStore on the same device.

[git-p4: depot-paths = "//depot/projects/EVALUATION/Main/": change = 7132]