Change Framework::Enumeration to be a collection of methods for each type, instead of a magical collection of templates and conversion operators.
This change is necessary because the static member enumToStringMap of
the template EnumerationFunctions could not be defined when the template
was instantiated. Although clang++/g++ were smart enough to figure it
out, clang++ recently started issuing very verbose warnings about this
problem. As an added benefit, the operators and methods are exposed to
the API more cleanly and they do not pollute the global namespace as
much.

To update code that is not under version control here:

 * In your .h:
   - #include <be_framework_enumeration.h>
   - In the global namespace, run this macro:
         BE_FRAMEWORK_ENUMERATION_DECLARATIONS(
             Fully qualified type,
             Name of std::map<enum, std::string> in your .cpp);

 * In your .cpp:
   - #include <file with BE_FRAMEWORK_ENUMERATION_DECLARATIONS>
   - In the global namespace, define a const std::map of your
     enumeration to string. 
   - In the global namespace, run this macro:
         BE_FRAMEWORK_ENUMERATION_DEFINITIONS(
             Fully qualified type,
             Name of std::map<enum, std::string> in your .cpp);

When you want to use methods provided by these macros, you'll find them
in the namespace BiometricEvaluation::Framework::Enumeration. It may be
convenient to issue a "using namespace" declaration for this namespace.

Refer to the files in this change list for examples.

[git-p4: depot-paths = "//depot/projects/EVALUATION/Main/": change = 8513]